backend_service/security.py
from __future__ import annotations

import hashlib
import hmac
import os
from datetime import UTC, datetime, timedelta

import jwt

ALGORITHM = "HS256"
PASSWORD_ALGORITHM = "pbkdf2_sha256"
PASSWORD_ITERATIONS = 600_000


def _jwt_secret() -> str:
    return os.getenv("AUTH_JWT_SECRET", "dev-auth-secret-change-me-32-bytes")


def hash_password(password: str, *, salt_hex: str | None = None) -> str:
    salt = bytes.fromhex(salt_hex) if salt_hex else os.urandom(16)
    digest = hashlib.pbkdf2_hmac(
        "sha256",
        password.encode("utf-8"),
        salt,
        PASSWORD_ITERATIONS,
    )
    return (
        f"{PASSWORD_ALGORITHM}${PASSWORD_ITERATIONS}$"
        f"{salt.hex()}${digest.hex()}"
    )


def verify_password(password: str, encoded_password: str) -> bool:
    try:
        algorithm, iterations_str, salt_hex, digest_hex = encoded_password.split("$", 3)
    except ValueError:
        return False

    if algorithm != PASSWORD_ALGORITHM:
        return False

    try:
        iterations = int(iterations_str)
    except ValueError:
        return False

    digest = hashlib.pbkdf2_hmac(
        "sha256",
        password.encode("utf-8"),
        bytes.fromhex(salt_hex),
        iterations,
    )
    return hmac.compare_digest(digest.hex(), digest_hex)


def create_access_token(*, user_id: str, email: str, role: str) -> str:
    ttl_hours = int(os.getenv("AUTH_TOKEN_TTL_HOURS", "24"))
    now = datetime.now(UTC)
    payload = {
        "sub": user_id,
        "email": email,
        "role": role,
        "iat": int(now.timestamp()),
        "exp": int((now + timedelta(hours=ttl_hours)).timestamp()),
    }
    return jwt.encode(payload, _jwt_secret(), algorithm=ALGORITHM)


def decode_access_token(token: str) -> dict[str, str]:
    payload = jwt.decode(token, _jwt_secret(), algorithms=[ALGORITHM])
    return {
        "user_id": str(payload["sub"]),
        "email": str(payload["email"]),
        "role": str(payload["role"]),
    }
